Auth module done

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { InjectRepository } from '@nestjs/typeorm/dist';
 import { TaskRepository } from './tasks.repository';
 import { Task } from './tasks.entity';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -13,52 +14,32 @@ export class TasksService {
         @InjectRepository(TaskRepository)
         private taskRepository: TaskRepository){}
 
-    // getAllTasks(): Task[] {
-    //     return this.tasks
-    // }
-
-    // getTasksWithFilter(filterDto:GetTasksFilterDto):Task[] {
-    //     const {status, search} = filterDto
-    //     let tasks = this.getAllTasks()
-    //     if(status){
-    //         tasks = tasks.filter((task) => task.status === status)
-    //     }
-        
-    //     if(search){
-    //         tasks = tasks.filter(task=>
-    //             task.title.includes(search) ||
-    //             task.description.includes(search))
-    //     }
-
-    //     return tasks
-    // }
-
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]>{
-        return this.taskRepository.getTasks(filterDto)
+    async getTasks(filterDto: GetTasksFilterDto,user: User): Promise<Task[]>{
+        return this.taskRepository.getTasks(filterDto,user)
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task>{
-        return this.taskRepository.createTask(createTaskDto)
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task>{
+        return this.taskRepository.createTask(createTaskDto,user)
     }
 
-    async getTaskById(id: number): Promise<Task>{
-       const found= await this.taskRepository.findOne(id)
+    async getTaskById(id: number, user: User): Promise<Task>{
+       const found= await this.taskRepository.findOne({where:{ id, userId: user.id }})
        if(!found){
            throw new NotFoundException(`Task with ID: "${id}" not found`)  
        }
        return found
     }
 
-    async deleteTaskById(id: number): Promise<void>{
-       const result = await this.taskRepository.delete(id)
-      
-       if(result.affected ===0) {
+    async deleteTaskById(id: number,user: User): Promise<void>{
+       const result = await this.taskRepository.delete({id, userId: user.id})
+       
+       if(result.affected === 0) {
             throw new NotFoundException(`Task with ID: "${id}" not found`)
        }
     }
 
-    async updateTaskStatus(id: number,status: TaskStatus): Promise<Task>{
-       const task = await this.getTaskById(id)
+    async updateTaskStatus(id: number,status: TaskStatus,user: User): Promise<Task>{
+       const task = await this.getTaskById(id,user)
        task.status = status
        await task.save()
        return task
